Revert "Private check"
