+++ src/app/component/[component]/layout.tsx
@@ -10,14 +10,10 @@ import { Tabs, Tab } from '@leafygreen-ui/tabs';
 import { spacing } from '@leafygreen-ui/tokens';
 import { H2 } from '@leafygreen-ui/typography';
 
-import { useSession } from '@/hooks';
 import { SubPath, getGithubLink } from '@/utils';
 import { useContentStackContext } from '@/contexts/ContentStackContext';
 
 import { titleCase } from '@/utils/titleCase';
-import { PrivateContentWall } from '@/components/global';
-
-import { NotFound } from '@/components/global/NotFound';
 
 const liveExamplePath = 'live-example';
 const designDocsPath = 'design-docs';
@@ -28,7 +24,6 @@ export default function ComponentLayout({
 }: {
   children: React.ReactNode;
 }) {
-  const { isLoggedIn } = useSession();
   const router = useRouter();
   const pathname = usePathname();
   const currentComponent = pathname.split('/')[2];
@@ -38,13 +33,6 @@ export default function ComponentLayout({
   // Find component in context. This will include the data from Contentstack
   const component = componentsFromContext.find(
     component => component.title === componentTitle,
-  );
-
-  if (!component) return <NotFound />;
-
-  const isComponentPrivate = component?.private;
-  const shouldRenderPrivateContentWall = Boolean(
-    isComponentPrivate && !isLoggedIn,
   );
 
   const getSelected = () => {
@@ -90,86 +78,77 @@ export default function ComponentLayout({
         min-height: 100vh;
       `}
     >
-      {shouldRenderPrivateContentWall ? (
-        <PrivateContentWall />
-      ) : (
-        <>
-          <H2
-            className={css`
-              text-transform: capitalize;
-              margin-bottom: ${spacing[600]}px;
-            `}
-          >
-            {currentComponent.split('-').join(' ')}
-          </H2>
+      <H2
+        className={css`
+          text-transform: capitalize;
+          margin-bottom: ${spacing[600]}px;
+        `}
+      >
+        {currentComponent.split('-').join(' ')}
+      </H2>
 
-          <Tabs
-            selected={getSelected()}
-            aria-label="main tabs"
-            className={css`
-              margin-bottom: ${spacing[800]}px;
-            `}
-            inlineChildren={
-              <>
-                {externalLinks.map(
-                  (
-                    { 'aria-label': ariaLabel, href, icon, isPrivate },
-                    index,
-                  ) => {
-                    if (isPrivate && !isLoggedIn) {
-                      return null;
-                    }
+      <Tabs
+        selected={getSelected()}
+        aria-label="main tabs"
+        className={css`
+          margin-bottom: ${spacing[800]}px;
+        `}
+        inlineChildren={
+          <>
+            {externalLinks.map(
+              ({ 'aria-label': ariaLabel, href, icon, isPrivate }, index) => {
+                if (isPrivate) {
+                  return null;
+                }
 
-                    if (!href) {
-                      return null;
-                    }
+                if (!href) {
+                  return null;
+                }
 
-                    return (
-                      <IconButton
-                        key={ariaLabel + index}
-                        aria-label={ariaLabel}
-                        size="large"
-                        href={href}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                      >
-                        {icon}
-                      </IconButton>
-                    );
-                  },
-                )}
-              </>
-            }
-          >
-            <Tab
-              onClick={() =>
-                router.push(`/component/${currentComponent}/${liveExamplePath}`)
-              }
-              name="Live Example"
-            >
-              <></>
-            </Tab>
-            <Tab
-              onClick={() =>
-                router.push(`/component/${currentComponent}/${designDocsPath}`)
-              }
-              name="Design Documentation"
-            >
-              <></>
-            </Tab>
-            <Tab
-              onClick={() =>
-                router.push(`/component/${currentComponent}/${codeDocsPath}`)
-              }
-              name="Code Documentation"
-            >
-              <></>
-            </Tab>
-          </Tabs>
+                return (
+                  <IconButton
+                    key={ariaLabel + index}
+                    aria-label={ariaLabel}
+                    size="large"
+                    href={href}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                  >
+                    {icon}
+                  </IconButton>
+                );
+              },
+            )}
+          </>
+        }
+      >
+        <Tab
+          onClick={() =>
+            router.push(`/component/${currentComponent}/${liveExamplePath}`)
+          }
+          name="Live Example"
+        >
+          <></>
+        </Tab>
+        <Tab
+          onClick={() =>
+            router.push(`/component/${currentComponent}/${designDocsPath}`)
+          }
+          name="Design Documentation"
+        >
+          <></>
+        </Tab>
+        <Tab
+          onClick={() =>
+            router.push(`/component/${currentComponent}/${codeDocsPath}`)
+          }
+          name="Code Documentation"
+        >
+          <></>
+        </Tab>
+      </Tabs>
 
-          <div>{children}</div>
-        </>
-      )}
+      <div>{children}</div>
     </div>
   );
 }
