+++ src/app/layout.tsx
@@ -1,11 +1,9 @@
 'use server';
 
 import './globals.css';
-import { auth } from '@/auth';
-import { SessionProvider } from 'next-auth/react';
 
 import LayoutWrapper from '@/components/layout-wrapper';
-import { fetchComponentsService } from '@/lib/contentStack/contentStackService';
+import { getComponentsService } from '@/lib/contentStack/contentStackService';
 
 export default async function RootLayout({
   children,
@@ -13,15 +11,7 @@ export default async function RootLayout({
   children: React.ReactNode;
 }>) {
   // This can directly call the ContentStack SDK to fetch the content page since this is a server component
+  const components = await getComponentsService({ includeContent: false });
 
-  const [session, components] = await Promise.all([
-    auth(),
-    fetchComponentsService({ includeContent: false }),
-  ]);
-
-  return (
-    <SessionProvider session={session}>
-      <LayoutWrapper components={components}>{children}</LayoutWrapper>
-    </SessionProvider>
-  );
+  return <LayoutWrapper components={components}>{children}</LayoutWrapper>;
 }
