+++ src/app/private/page.tsx
@@ -5,17 +5,13 @@ import Button from '@leafygreen-ui/button';
 import { BasicEmptyState } from '@leafygreen-ui/empty-state';
 // @ts-expect-error
 import ArrowLeftIcon from '@leafygreen-ui/icon/dist/ArrowLeft';
-// @ts-expect-error
-import LogInIcon from '@leafygreen-ui/icon/dist/LogIn';
-import { login } from '@/auth';
-import { ComingSoon, Security } from '@/components/glyphs';
-import { useSession } from '@/hooks';
+
+import { ComingSoon } from '@/components/glyphs';
 
 export default function Private() {
   const router = useRouter();
-  const { isLoggedIn } = useSession();
 
-  return isLoggedIn ? (
+  return (
     <BasicEmptyState
       title="Coming Soon"
       description="Check back for updates soon"
@@ -25,21 +21,6 @@ export default function Private() {
         </Button>
       }
       graphic={<ComingSoon />}
-    />
-  ) : (
-    <BasicEmptyState
-      title="Log in to view private content"
-      description="This page is locked for security purposes and only accessible by MongoDB employees."
-      primaryButton={
-        <Button
-          variant="primary"
-          onClick={() => login()}
-          leftGlyph={<LogInIcon />}
-        >
-          Log In
-        </Button>
-      }
-      graphic={<Security />}
     />
   );
 }
