+++ src/components/global/SideNavigation/SideNavigation.tsx
@@ -24,7 +24,7 @@ import {
 import { MongoDBLogo, SupportedColors } from '@leafygreen-ui/logo';
 import { color, spacing } from '@leafygreen-ui/tokens';
 import { SIDE_NAV_WIDTH } from '@/constants';
-import { useMediaQuery, useSession } from '@/hooks';
+import { useMediaQuery } from '@/hooks';
 import { components } from '@/utils/components';
 import { foundations } from '@/utils/foundations';
 import type { SubPathMeta } from '@/utils/types';
@@ -36,10 +36,7 @@ import { SideNavLabel } from './SideNavLabel';
 import { SideNavList } from './SideNavList';
 import { shouldAddColonToTitle } from '@/utils';
 
-import { PrivateIcon as PrivateIconComponent } from '@/components/global/PrivateIcon';
-
 export function SideNavigation() {
-  const { isLoggedIn } = useSession();
   const navRef = useRef<HTMLElement>(null);
   const [open, setOpen] = React.useState(false);
   const [isMobile] = useMediaQuery(['(max-width: 640px)'], {
@@ -57,15 +54,6 @@ export function SideNavigation() {
           resource.name.toLowerCase()
       : pathname === resource.navPath;
   };
-
-  const PrivateIcon = () => (
-    <PrivateIconComponent
-      isPrivate={!isLoggedIn}
-      className={css`
-        margin-left: ${spacing[400]}px;
-      `}
-    />
-  );
 
   const navContent = (
     <>
@@ -88,7 +76,6 @@ export function SideNavigation() {
             active={isActiveResource(foundation)}
           >
             {shouldAddColonToTitle(foundation.name)}
-            {foundation.isPrivate && <PrivateIcon />}
           </SideNavItem>
         ))}
       </SideNavList>
@@ -113,7 +100,6 @@ export function SideNavigation() {
             active={isActiveResource(pattern)}
           >
             {shouldAddColonToTitle(pattern.name)}
-            {pattern.isPrivate && <PrivateIcon />}
           </SideNavItem>
         ))}
       </SideNavList>
@@ -141,7 +127,6 @@ export function SideNavigation() {
             }
           >
             {component.name}
-            {component.isPrivate && <PrivateIcon />}
           </SideNavItem>
         ))}
       </SideNavList>
