+++ src/components/global/UserMenu.tsx
@@ -58,4 +58,6 @@ export function UserMenu() {
   ) : (
     <LogIn />
   );
+
+  return null;
 }
