+++ src/utils/components.ts
@@ -89,11 +89,6 @@ export const components: Array<SubPathMeta> = [
     navPath: generateComponentNavPath(ComponentSubPath.Callout),
   },
   {
-    name: titleCase(ComponentSubPath.CanvasHeader),
-    navPath: generateComponentNavPath(ComponentSubPath.CanvasHeader),
-    isPrivate: true,
-  },
-  {
     name: titleCase(ComponentSubPath.Checkbox),
     navPath: generateComponentNavPath(ComponentSubPath.Checkbox),
   },
